refactor: move validation + steam integration logic into LocationsTabController
